矩阵等实现代码

+++ app/MyGraphicsCode/MyGraphicsCode/MathUtil.cpp
@@ -20,5 +20,5 @@ float safeAcos(float x)
 		return 0.0f;
 	}
 
-	return acos(x);
+	return acosf(x);
 }
